requirements.txt
asgiref==3.3.1
autopep8==1.5.5
certifi==2020.12.5
chardet==4.0.0
Django==3.1.6
django-cors-headers==3.7.0
djangorestframework==3.12.2
idna==2.10
psycopg2-binary==2.8.6
pycodestyle==2.6.0
pytz==2021.1
requests==2.25.1
sqlparse==0.4.1
toml==0.10.2
urllib3==1.26.3
